+++ .gitignore
@@ -15,6 +15,8 @@ _native.cpython*
 .vscode/
 dev/
 pkg/
+# Staged temporarily by `rum wasm build` so wasm-pack can include crate-local license metadata.
+crates/rumoca-bind-wasm/LICENSE
 perf.data*
 flamegraph.svg
 msl
