+++ crates/rumoca-bind-wasm/Cargo.toml
@@ -23,6 +23,7 @@ serde_json = { workspace = true }
 lsp-types = { workspace = true }
 serde-wasm-bindgen = "0.6"
 js-sys = "0.3"
+wasm-bindgen-futures = "0.4"
 console_error_panic_hook = { version = "0.1", optional = true }
 wasm-bindgen-rayon = { version = "1.2", optional = true }
